Fix final callbacks for reentrant backwards
Closes #24965

Prior to this commit, final_callbacks_ are cleared on exit of ANY
backward. When using reentrant backward, the last backward would
remove all callbacks from the engine. However, this might lead to
unexpected behavior. For example, the application could install
a final callback after forward, and expecting this callback to fire
when all gradients are ready. If there is a renentrant backward on
a subgraph, it would fire the callback and delete it on exit,
meaning that when fired, not all gradients are ready.

**Failed Attempt**

The 1st attempt was trying to move the callback to the GraphTask
in engine::execute(). However, this failed because more callbacks
could be installed during backward pass.

**Current Solution**

Final callbacks are stored as part of the context for each reentrant
backward. Thisi PR converts final_callbacks_ into a map that maps
reentrant depth to the callbacks installed during the lifetime of the
corresponding backward. More specifically:

* Insertion: when inserting callbacks, queue_callback() uses the
  thread_local total_depth to find the vector to insert into.
* Deletion: the ClearCallbacks guard remembers the depth of the current
  backward call, and only erases the corresponding vector from the map
  on exit.
* Execution: at the end of backward with depth x, the engine only
  executes callbacks installed during the corresponding backward with
  depth x.

Note that: Using a map instead of a vector to avoid the overhead of
creating an empty vector for each depth, while the callbacks might be
only installed in a small subset of those reentrant backward calls.

ghstack-source-id: 9449db06fc9aaa87930d1716d43fbf2b97a29179
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35066